clients/customize: rename profilepage to public page
